refactor: Position 클래스에서 최대 이동 가능 거리 계산 방법 수정
